backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
whisper==1.1.10
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
pillow==10.1.0
openai==1.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 